pyproject.toml
[project]
name = "raytracer"
version = "0.1.0"
description = "A raytracer for experimenting with rendering logic"
requires-python = ">=3.11"
dependencies = [
    "taichi>=1.7.0",
    "numpy>=1.26.0",
    "Pillow>=10.0.0",
    "matplotlib>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

# Taichi uses its own DSL with decorators and types that mypy cannot understand.
# We disable specific errors for modules where Taichi is heavily used.
[[tool.mypy.overrides]]
module = "src.python.core.*"
disable_error_code = ["valid-type", "no-any-return", "call-arg", "attr-defined", "no-untyped-def", "import-untyped", "operator", "no-untyped-call"]
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = "src.python.geometry.*"
disable_error_code = ["valid-type", "no-any-return", "call-arg", "attr-defined", "no-untyped-def", "import-untyped", "operator"]
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = "src.python.scene.*"
disable_error_code = ["valid-type", "no-any-return", "call-arg", "attr-defined", "no-untyped-def", "import-untyped", "operator"]
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = "src.python.materials.*"
disable_error_code = ["valid-type", "no-any-return", "call-arg", "attr-defined", "no-untyped-def", "import-untyped", "operator"]
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = "src.python.camera.*"
disable_error_code = ["valid-type", "no-any-return", "call-arg", "attr-defined", "no-untyped-def", "import-untyped", "operator"]
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = "src.python.preview.*"
disable_error_code = ["no-any-return"]

[[tool.mypy.overrides]]
module = "tests.*"
disable_error_code = ["no-untyped-def", "import-untyped", "call-arg", "untyped-decorator", "arg-type", "no-untyped-call"]

[[tool.mypy.overrides]]
module = "examples.*"
disable_error_code = ["import-untyped"]

[tool.pytest.ini_options]
testpaths = ["src/python", "tests"]
python_files = ["*_test.py", "test_*.py"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
